requirements.txt
torch==2.5.1
typer==0.15.1
fastapi==0.115.6
uvicorn==0.34.0
pillow==11.1.0
numpy==2.2.1
torchvision==0.20.1
kagglehub==0.3.6
dvc==3.59.0
pytorch_lightning==2.5.0.post0
matplotlib==3.10.0
google-api-python-client==2.159
dvc-gs==3.0.1
python-multipart==0.0.20
uvicorn[standard]==0.34.0
hydra-core==1.3.2
prometheus_client==0.21.1
